requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
docker==7.1.0
docutils==0.21.2
drf-spectacular==0.27.2
flake8==7.1.1
flake8-docstrings==1.7.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
snowballstemmer==2.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==2.3.0
